+++ src/umka_gen.c
@@ -376,7 +376,7 @@ void genPushRealConst(CodeGen *gen, double realVal)
 
 void genPushGlobalPtr(CodeGen *gen, void *ptrVal)
 {
-    const Instruction instr = {.opcode = OP_PUSH, .tokKind = TOK_NONE, .typeKind = TYPE_PTR, .operand.ptrVal = (int64_t)ptrVal};
+    const Instruction instr = {.opcode = OP_PUSH, .tokKind = TOK_NONE, .typeKind = TYPE_PTR, .operand.ptrVal = ptrVal};
     genAddInstr(gen, &instr);
 }
 
@@ -481,7 +481,7 @@ void genChangeRefCnt(CodeGen *gen, TokenKind tokKind, Type *type)
 {
     if (typeGarbageCollected(type))
     {
-        const Instruction instr = {.opcode = OP_CHANGE_REF_CNT, .tokKind = tokKind, .typeKind = TYPE_NONE, .operand.ptrVal = (int64_t)type};
+        const Instruction instr = {.opcode = OP_CHANGE_REF_CNT, .tokKind = tokKind, .typeKind = TYPE_NONE, .operand.ptrVal = type};
         genAddInstr(gen, &instr);
     }
 }
@@ -491,7 +491,7 @@ void genChangeRefCntAssign(CodeGen *gen, Type *type)
 {
     if (typeGarbageCollected(type))
     {
-        const Instruction instr = {.opcode = OP_CHANGE_REF_CNT_ASSIGN, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = (int64_t)type};
+        const Instruction instr = {.opcode = OP_CHANGE_REF_CNT_ASSIGN, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = type};
         genAddInstr(gen, &instr);
     }
     else
@@ -503,7 +503,7 @@ void genSwapChangeRefCntAssign(CodeGen *gen, Type *type)
 {
     if (typeGarbageCollected(type))
     {
-        const Instruction instr = {.opcode = OP_CHANGE_REF_CNT_ASSIGN, .inlineOpcode = OP_SWAP, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = (int64_t)type};
+        const Instruction instr = {.opcode = OP_CHANGE_REF_CNT_ASSIGN, .inlineOpcode = OP_SWAP, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = type};
         genAddInstr(gen, &instr);
     }
     else
@@ -560,7 +560,7 @@ void genGetFieldPtr(CodeGen *gen, int fieldOffset)
 
 void genAssertType(CodeGen *gen, Type *type)
 {
-    const Instruction instr = {.opcode = OP_ASSERT_TYPE, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = (int64_t)type};
+    const Instruction instr = {.opcode = OP_ASSERT_TYPE, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = type};
     genAddInstr(gen, &instr);
 }
 
@@ -616,7 +616,7 @@ void genCallIndirect(CodeGen *gen, int paramSlots)
 
 void genCallExtern(CodeGen *gen, void *entry)
 {
-    const Instruction instr = {.opcode = OP_CALL_EXTERN, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = (int64_t)entry};
+    const Instruction instr = {.opcode = OP_CALL_EXTERN, .tokKind = TOK_NONE, .typeKind = TYPE_NONE, .operand.ptrVal = entry};
     genAddInstr(gen, &instr);
 }
